Export of internal ZetaSQL changes.
Add support for read-modify-write of java ResolvedAST (see RewritingVisitorTest.java for examples)
Add support for TO_JSON_STRING.

--
Change by ZetaSQL Team <no-reply@google.com>:
Pass correct language options to Validator in PreparedQuery api when validating a resolved tree passed to the PreparedQuery constructor.
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove in_development tag from FEATURE_BIGNUMERIC_TYPE.
--
Change by ZetaSQL Team <no-reply@google.com>:
Increase the number of operations per benchmark loop in value_test
--
Change by ZetaSQL Team <no-reply@google.com>:
Update zetasql documentation to include BIGNUMERIC support.
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove unused lexer rules.
--
Change by ZetaSQL Team <no-reply@google.com>:
Analyzer changes for QUALIFY clause
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix DEFAULT column syntax error messages.
--
Change by ZetaSQL Team <no-reply@google.com>:
Make a clarification about the promotion of NULLs in GREATEST/LEAST
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix bug with MODIFY_MAP rewriter - NULL first arg should be NULL output
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix bug in handling array offsets with FLATTEN.
--
Change by ZetaSQL Team <no-reply@google.com>:
execute_query: Add JSON and Textproto output modes
--
Change by Matthew Brown <matthewbr@google.com>:
Fix name and contract for DescriptorPool.getAllFileDescriptors to be ordered
--
Change by ZetaSQL Team <no-reply@google.com>:
remove expected error for contains_key
--
Change by ZetaSQL Team <no-reply@google.com>:
Undeprecate presence tests in proto3.
--
Change by ZetaSQL Team <no-reply@google.com>:
INTERNAL
--
Change by ZetaSQL Team <no-reply@google.com>:
execute_query: Move "ExecuteQuery" overload into test code
--
Change by ZetaSQL Team <no-reply@google.com>:
execute_query: Retrieve examination callback only once
--
Change by ZetaSQL Team <no-reply@google.com>:
execute_query: Add writers for Textproto and JSON formats
--
Change by Matthew Brown <matthewbr@google.com>:
local_service: Use a more explicit for serializing DescriptorPools
--
Change by ZetaSQL Team <no-reply@google.com>:
Support parameterized types in MERGE INSERT
--
Change by ZetaSQL Team <no-reply@google.com>:
Add utilities to generate protobuf messages from iterator
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix resolver logic for PIVOT to handle the case where columns in the pivot input have multiple aliases, but only some of them are referenced.
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove deprecated AddCastOrConvertLiteral function
--
Change by Matthew Brown <matthewbr@google.com>:
Remove Serializable from DescriptorPool.sub interfaces
--
Change by ZetaSQL Team <no-reply@google.com>:
Add explicit CAST node to struct fields with type parameters
--
Change by ZetaSQL Team <no-reply@google.com>:
Change the visibility of cast_date_time.
--
Change by ZetaSQL Team <no-reply@google.com>:
Use separate control flow nodes for starting a FOR loop and advancing a FOR loop.
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix the AST locations for TVF arguments. If it refers to an expression, just pass the location of the expression to the downstream. Otherwise it would prevent literal casting for struct fields.
--
Change by ZetaSQL Team <no-reply@google.com>:
Use null handling behavior enum in reference impl instead of hard-coded list of functions that respect nulls.
--
Change by ZetaSQL Team <no-reply@google.com>:
Modify TypeParameters:MatchType()
--
Change by ZetaSQL Team <no-reply@google.com>:
Automatically FLATTEN for IN UNNEST.
--
Change by ZetaSQL Team <no-reply@google.com>:
Native PIVOT implementation in reference impl, which doesn't have limitations on types of aggregate functions supported imposed by the rewriter.
--
Change by ZetaSQL Team <no-reply@google.com>:
Split long logging statements to multiple lines.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add flag to analyze_query to control rewrites.
--
Change by ZetaSQL Team <no-reply@google.com>:
RQG for ARRAY_FILTER and ARRAY_TRANSFORM.
--
Change by ZetaSQL Team <no-reply@google.com>:
execute_query: Add option to gracefully handle statement prompt errors
--
Change by ZetaSQL Team <no-reply@google.com>:
Add string escaping in test_function.cc.
  (And 84 more changes)

GitOrigin-RevId: de685b7ab6095676abb23b49ff0dfc3e51eb09d5

+++ zetasql/public/functions/net.cc
@@ -21,6 +21,8 @@
 #include <string.h>
 #include <sys/socket.h>
 
+#include <cstdint>
+
 #include "zetasql/public/functions/util.h"
 #include "absl/base/optimization.h"
 #include "absl/strings/ascii.h"
@@ -228,7 +230,7 @@ bool IPTrunc(absl::string_view in, int64_t prefix_length_bits, std::string* out,
   const uint8_t remaining_bits = prefix_length_bits % 8;
   if (remaining_bits > 0) {
     *out_data = static_cast<uint8_t>(in[bytes_to_keep]) &
-        static_cast<uint8_t>(0xff00u >> remaining_bits);
+                static_cast<uint8_t>(0xff00u >> remaining_bits);
   }
   return true;
 }
@@ -555,7 +557,7 @@ absl::Status DomainSuffix(
   std::string normalized_host_storage;
   absl::string_view normalized_host = host;
   for (const char c : host) {
-    if (kHostCharMap[c] & kNormalize) {
+    if (kHostCharMap[static_cast<uint8_t>(c)] & kNormalize) {
       if (!zetasql::internal::ToASCII(host, &normalized_host_storage)) {
         return absl::OkStatus();
       }
